@@ -12,48 +12,46 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.3",
-    "@angular/common": "^6.1.3",
-    "@angular/compiler": "^6.1.3",
-    "@angular/core": "^6.1.3",
-    "@angular/forms": "^6.1.3",
-    "@angular/http": "^6.1.3",
-    "@angular/platform-browser": "^6.1.3",
-    "@angular/platform-browser-dynamic": "^6.1.3",
-    "@angular/pwa": "^0.7.4",
-    "@angular/router": "^6.1.3",
-    "@angular/service-worker": "^6.1.3",
-    "@ngtools/webpack": "^6.1.4",
+  "@angular/animations": "~7.1.0",
+    "@angular/common": "~7.1.0",
+    "@angular/compiler": "~7.1.0",
+    "@angular/core": "~7.1.0",
+    "@angular/forms": "~7.1.0",
+    "@angular/platform-browser": "~7.1.0",
+    "@angular/platform-browser-dynamic": "~7.1.0",
+    "@angular/pwa": "^0.7.5",
+    "@angular/router": "~7.1.0",
+    "@angular/service-worker": "^7.1.1",
     "@ngx-translate/core": "^10.0.2",
     "@ngx-translate/http-loader": "^3.0.1",
-    "core-js": "^2.5.7",
+    "core-js": "^2.5.4",
     "http-server": "^0.11.1",
-    "nodemailer": "^4.6.8",
-    "rxjs": "^6.2.2",
-    "rxjs-compat": "^6.2.2",
-    "sweetalert": "^2.1.0",
+    "nodemailer": "^4.7.0",
+    "rxjs": "^6.3.3",
+    "tslib": "^1.9.0",
+    "sweetalert": "^2.1.2",
     "tablesaw": "^3.0.9",
-    "zone.js": "^0.8.26"
+    "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.7.4",
-    "@angular/cli": "^6.2.3",
-    "@angular/compiler-cli": "^6.1.3",
-    "@angular/language-service": "^6.1.3",
-    "@types/jasmine": "^2.8.8",
+    "@angular-devkit/build-angular": "~0.11.0",
+    "@angular/cli": "~7.1.0",
+    "@angular/compiler-cli": "~7.1.0",
+    "@angular/language-service": "~7.1.0",
+    "@types/node": "~8.9.4",
+    "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~10.7.1",
-    "codelyzer": "~4.4.4",
-    "jasmine-core": "~3.2.1",
+    "codelyzer": "~4.5.0",
+    "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^3.0.0",
+    "karma": "~3.1.1",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^2.0.1",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^1.3.1",
-    "protractor": "^5.4.0",
-    "ts-node": "~7.0.1",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~1.1.2",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "^2.9.2"
+    "typescript": "~3.1.6"
   }
 }
