@@ -3,6 +3,9 @@ import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { URL_SERVICIOS } from '../../config/config';
 import { Router } from '@angular/router';
+import swal from 'sweetalert';
+import { Observable, throwError } from 'rxjs';
+import { map, catchError} from 'rxjs/operators';
 
 @Injectable()
 export class EmailService {
@@ -15,11 +18,12 @@ export class EmailService {
      // tslint:disable-next-line:prefer-const
      let url = URL_SERVICIOS + '/formulario';
      return this.http.post(url, email)
-     .map((resp: any) => {
+     .pipe(
+      map((resp: any) => {
       swal('Correo Enviado!', email.nombre , 'success');
        return resp.email;
 
-     });
+     }));
 
    }
 }
