@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { Router, ActivationEnd } from '@angular/router';
 import { Title, Meta, MetaDefinition } from '@angular/platform-browser';
+import { Observable, throwError } from 'rxjs';
+import { map, catchError, filter } from 'rxjs/operators';
 
 @Component({
   selector: 'app-breadcrumbs',
@@ -35,8 +37,9 @@ this.getDataRouter()
   }
 getDataRouter() {
     return this.router.events
-    .filter( evento => evento instanceof ActivationEnd )
-    .filter( (evento: ActivationEnd) => evento.snapshot.firstChild === null )
-    .map( (evento: ActivationEnd) => evento.snapshot.data);
+    .pipe(
+    filter( evento => evento instanceof ActivationEnd ),
+    filter( (evento: ActivationEnd) => evento.snapshot.firstChild === null ),
+    map( (evento: ActivationEnd) => evento.snapshot.data));
 }
  }
